requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
mangum==0.17.0
aiogram==3.13.1
boto3==1.35.0
httpx==0.27.2
requests==2.32.3
pillow==11.1.0
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.12
